requirements.txt
flask==3.1.0
flask-smorest==0.45.0
python-dotenv==1.0.1
sqlalchemy==2.0.39
flask_sqlalchemy==3.1.1
flask_jwt_extended==4.7.1
passlib==1.7.4
flask-migrate==4.1.0
gunicorn==23.0.0
psycopg2-binary==2.9.10
requests==2.32.3
rq==2.3.1
redis==5.2.1